requirements.txt
asgiref==3.7.2
Brotli==1.0.9
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.0.0
Django==4.2.2
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.3
tzdata==2023.3
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
